Corrigindo o tipo das variáveis.

+++ Desafio35.py
@@ -1,7 +1,7 @@
 print("VERIFIQUE SE TRÊS RETAS FORMAM  UM TRIÂNGULO!")
-reta_X = int(input("Digite a primeiro reta: "))
-reta_Y = int(input("Digite a segundo reta: "))
-reta_Z = int(input("Digite a terceiro reta: "))
+reta_X = float(input("Digite a primeiro reta: "))
+reta_Y = float(input("Digite a segundo reta: "))
+reta_Z = float(input("Digite a terceiro reta: "))
 
 if reta_X + reta_Y > reta_Z and reta_Y + reta_Z > reta_X and reta_Z + reta_X > reta_Y:
     print("O triangulo pode ser formado!")
